Update my project files and add a README  content
